requirements.txt
streamlit==1.31.0
pandas==2.0.3
numpy==1.24.3
plotly==5.18.0
prophet==1.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.3
seaborn==0.13.0
